moving seperate service into current coinbase service
